requirements.txt
openai>=1.0.0
pypdf>=5.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
